# schema/api.yaml
openapi: 3.0.3
info:
  title: Arch Blog API
  description: A modern blog platform backend API
  version: 1.0.0
  contact:
    name: API Support
    email: support@archblog.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.archblog.com/v1
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: object
          additionalProperties: true
...
    PaginationMeta:
      type: object
      required:
        - totalItems
        - totalPages
        - currentPage
        - itemsPerPage
      properties:
        totalItems:
          type: integer
          minimum: 0
          example: 100
        totalPages:
          type: integer
          minimum: 0
          example: 10
        currentPage:
          type: integer
          minimum: 1
          example: 1
        itemsPerPage:
          type: integer
          minimum: 1
          maximum: 100
          example: 10
...
    Post:
      type: object
      required:
        - id
        - title
        - content
        - excerpt
        - slug
        - status
        - authorId
        - viewCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: "Introduction to Hexagonal Architecture"
        content:
          type: string
          minLength: 1
          example: "<p>This post explains the principles of hexagonal architecture...</p>"
        excerpt:
          type: string
          maxLength: 500
          example: "A comprehensive guide to understanding hexagonal architecture"
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          maxLength: 255
          example: "introduction-to-hexagonal-architecture"
        status:
          type: string
          enum: [draft, published, archived]
          example: "published"
        authorId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        viewCount:
          type: integer
          minimum: 0
          example: 1234
        publishedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    PostSummary:
      type: object
      required:
        - id
        - title
        - excerpt
        - slug
        - status
        - authorId
        - viewCount
        - createdAt
        - publishedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Introduction to Hexagonal Architecture"
        excerpt:
          type: string
          example: "A comprehensive guide to understanding hexagonal architecture"
        slug:
          type: string
          example: "introduction-to-hexagonal-architecture"
        status:
          type: string
          enum: [draft, published, archived]
          example: "published"
        authorId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        viewCount:
          type: integer
          minimum: 0
          example: 1234
        publishedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    CreatePostRequest:
      type: object
      required:
        - title
        - content
        - excerpt
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: "Introduction to Hexagonal Architecture"
        content:
          type: string
          minLength: 1
          example: "<p>This post explains the principles of hexagonal architecture...</p>"
          description: "HTML content of the post (will be sanitized)"
        excerpt:
          type: string
          maxLength: 500
          example: "A comprehensive guide to understanding hexagonal architecture"
    UpdatePostRequest:
      type: object
      required:
        - title
        - content
        - excerpt
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: "Updated: Introduction to Hexagonal Architecture"
        content:
          type: string
          minLength: 1
          example: "<p>This updated post explains...</p>"
        excerpt:
          type: string
          maxLength: 500
          example: "An updated guide to hexagonal architecture"
    PaginatedPosts:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PostSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Themes schemas
...
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Missing or invalid authentication token"
    ForbiddenError:
      description: User lacks the required permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "You do not have permission to perform this action"
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource not found"
    ValidationError:
      description: The request data is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input data"
            details:
              username: "Username must be between 3 and 30 characters"
    ConflictError:
      description: The request conflicts with existing data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "conflict"
            message: "Username already exists"
    InternalServerError:
      description: An unexpected error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "internal_server_error"
            message: "An unexpected error occurred"
paths:
...
  /posts:
    get:
      tags:
        - Posts
      summary: List posts
      description: Returns a paginated list of posts
      operationId: listPosts
      security: []  # Public endpoint
      parameters:
        - name: status
          in: query
          description: Filter by post status
          schema:
            type: string
            enum: [draft, published, archived]
        - name: authorId
          in: query
          description: Filter by author ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [created_at, updated_at, published_at, title, view_count]
            default: created_at
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of posts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPosts'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Posts
      summary: Create a new post
      description: Creates a new blog post (initially in draft status)
      operationId: createPost
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Get a post by ID
      description: Returns a single post
      operationId: getPost
      security: []  # Public endpoint
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Posts
      summary: Update a post
      description: Updates an existing post
      operationId: updatePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Posts
      summary: Delete a post
      description: Permanently deletes a post
      operationId: deletePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Post deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /posts/slug/{slug}:
    get:
      tags:
        - Posts
      summary: Get a post by slug
      description: Returns a single post by its URL slug
      operationId: getPostBySlug
      security: []  # Public endpoint
      parameters:
        - name: slug
          in: path
          required: true
          description: The URL slug of the post
          schema:
            type: string
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      responses:
        '200':
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /posts/{id}/publish:
    post:
      tags:
        - Posts
      summary: Publish a post
      description: Transitions a post from draft to published status
      operationId: publishPost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post to publish
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /posts/{id}/unpublish:
    post:
      tags:
        - Posts
      summary: Unpublish a post
      description: Transitions a post back to draft status
      operationId: unpublishPost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post to unpublish
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post unpublished successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /posts/{id}/archive:
    post:
      tags:
        - Posts
      summary: Archive a post
      description: Transitions a post to archived status
      operationId: archivePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post to archive
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Themes endpoints
...